backend/requirements.txt
# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.14.0
pgvector==0.3.6

# Authentication
python-jose[cryptography]==3.3.0
python-multipart==0.0.12

# AI/ML - LangGraph + LangChain (compatible versions)
anthropic==0.39.0
openai==1.54.3
langgraph==0.2.55
langchain==0.3.9
langchain-core==0.3.21
langchain-community==0.3.9
langchain-anthropic==0.3.0
langsmith==0.1.147

# Local Embeddings (no API key needed)
sentence-transformers==3.3.1
torch==2.5.1
transformers==4.47.1

# Web Search
tavily-python==0.5.0

# PDF Processing
pypdf==5.1.0

# Microsoft Graph API for OneDrive
msal==1.31.0
msgraph-core==1.2.0
msgraph-sdk==1.12.0

# Environment & Config
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.10.1
email-validator==2.2.0
typing-extensions==4.12.2

# HTTP client
httpx==0.27.2

# Web scraping (for UVA resources)
beautifulsoup4==4.12.3
lxml==5.2.1
selenium==4.27.1

# Development (optional)
pytest==8.3.4
